Break classes up into submodules
