--- src/__mocks__/packageConfig.js
@@ -1,15 +1,0 @@
-let mockConfig = {};
-
-function setConfig(config = {}) {
-  mockConfig = config;
-}
-
-function getConfig() {
-  return mockConfig;
-}
-
-function requireFromString() {
-  return {};
-}
-
-module.exports = { getConfig, setConfig, requireFromString };
