+++ src/config/rc.js
@@ -1,0 +1,3 @@
+import PackageConfig from './package';
+
+export default class RcConfig extends PackageConfig {}
